app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
pydantic==2.6.3
gunicorn==21.2.0
boto3==1.34.69
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 